+++ code/drasil-lang/Language/Drasil/Development.hs
@@ -18,10 +18,10 @@ module Language.Drasil.Development (
   ) where
 
 import Language.Drasil.NounPhrase (NounPhrase(phraseNP,pluralNP))
+import Language.Drasil.DisplayExpr (DisplayExpr(..))
 import Language.Drasil.Expr (UFuncB(..), UFuncVec(..)
   , ArithBinOp(..), BoolBinOp(..), EqBinOp(..)
   , LABinOp(..), OrdBinOp(..), VVVBinOp(..), VVNBinOp(..))
-import Language.Drasil.Expr.Display
 import Language.Drasil.Expr.Extract
 import Language.Drasil.Expr.Precedence (precA, precB, eprec)
 import Language.Drasil.Sentence.Extract (sdep, lnames, lnames')
